more logging functions added

+++ src/edu/luc/cs/comp473/project1/model/maintenance/Order.java
@@ -40,7 +40,7 @@ public class Order {
      * 
      * @return order number
      */
-    public double getOrderNum() {
+    public int getOrderNum() {
         return orderNum;
     }
